abstract out coalesceScopes
